pyproject.toml
[project]
name = "pydantic-wizard"
version = "0.1.0"
description = "Interactive wizard-style CLI for configuring Pydantic v2 models, with YAML serialization."
readme = "README.md"
authors = [{ name = "Gianluca Pagliara", email = "pagliara.gianluca@gmail.com" }]
license = { text = "MIT" }
requires-python = ">=3.13"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "pydantic>=2.0",
    "typer>=0.15",
    "rich>=13.0",
    "questionary>=2.0",
    "PyYAML>=6.0",
]

[project.urls]
Homepage = "https://github.com/gianlucapagliara/pydantic-wizard"
Documentation = "https://gianlucapagliara.github.io/pydantic-wizard/"
Repository = "https://github.com/gianlucapagliara/pydantic-wizard"
Issues = "https://github.com/gianlucapagliara/pydantic-wizard/issues"

[project.scripts]
pydantic-wizard = "pydantic_wizard.app:app"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=6.0",
    "pytest-timeout>=2.3",
    "mypy>=1.7.0",
    "ruff>=0.8.4",
    "pre-commit>=4.0",
    "types-PyYAML>=6.0",
]
docs = [
    "mkdocs>=1.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
strict = true
disallow_untyped_decorators = false
ignore_missing_imports = true
disable_error_code = ["misc"]
exclude = ["tests/.*"]

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "C901",  # Function complexity
    "E501",  # Line too long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
timeout = 30
testpaths = ["tests"]
